Issue:#I7RT23

+++ test/fuzztest/abilitystubdoabilitybackground_fuzzer/abilitystubdoabilitybackground_fuzzer.cpp
@@ -33,7 +33,7 @@ const std::u16string ABILITYMGR_INTERFACE_TOKEN = u"ohos.aafwk.AbilityManager";
 
 bool DoSomethingInterestingWithMyAPI(const char* data, size_t size)
 {
-    uint32_t code = AbilityManagerInterfaceCode::DO_ABILITY_BACKGROUND;
+    uint32_t code = static_cast<uint32_t>(AbilityManagerInterfaceCode::DO_ABILITY_BACKGROUND);
 
     MessageParcel parcel;
     parcel.WriteInterfaceToken(ABILITYMGR_INTERFACE_TOKEN);
