+++ test/fuzztest/abilitystubregisterwmshandler_fuzzer/abilitystubregisterwmshandler_fuzzer.cpp
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-#include "abilitystubmoveabilitytobackground_fuzzer.h"
+#include "abilitystubregisterwmshandler_fuzzer.h"
 
 #include <cstddef>
 #include <cstdint>
@@ -33,7 +33,7 @@ const std::u16string ABILITYMGR_INTERFACE_TOKEN = u"ohos.aafwk.AbilityManager";
 
 bool DoSomethingInterestingWithMyAPI(const char* data, size_t size)
 {
-    uint32_t code = AbilityManagerInterfaceCode::MOVE_ABILITY_TO_BACKGROUND;
+    uint32_t code = static_cast<uint32_t>(AbilityManagerInterfaceCode::MOVE_ABILITY_TO_BACKGROUND);
 
     MessageParcel parcel;
     parcel.WriteInterfaceToken(ABILITYMGR_INTERFACE_TOKEN);
